Disable atomic file updates on Windows
Utilizing atomic file updates (via Chef, e.g. templates) means needing
Administrator permissions to read the SACL from files. While we expect
chef-client to be run as Administrator or with equivlanet permissions, we
expect chef-dk (chef) to be run as a regular user. Even when logged in as
an Administrator, you need to be running your shell with elevated
privileges for atomic file updates to work.

Fixes #109.