Ensure file_staging_uses_destdir is true
This way we are likely to inherit file ACLs on Windows when we create the
tempfile, since we're avoiding the code that reads and sets the ACLs on
a file move from file_atomic_update.

Chef 12 defaults to file_staging_uses_destdir being true, but prior releases
did not. This ensures we keep this setting.