Include dependency graph in policyfile lock
Policyfile lock can now check deps in updated cookbooks (though only
"local dev" cookbooks are allowed to be modified). To facilitate this, a
copy of the dependency graph is stored in the lockfile. As of this
commit, storing the graph in the lockfile is only partially implemented,
it must still be integrated into the policyfile compiler.