Use chefdk's gemfile to locate gems
This allows us to appbundle cookstyle correctly. cookstyle is a special
case because it has a runtime dependency on rubocop. Therefore,
cookstyle cannot be added to the gemfile for pretty much any other app
that uses rubocop (in particular, test-kitchen uses finstyle which is a
rubocop wrapper that conflicts with cookstyle). So cookstyle has to be
omitted from the shared gemfile that we edit into other apps' gemfiles.
However, you cannot `bundle show` a gem if that gem isn't part of your
bundle. So we have to use ChefDK's Gemfile (where cookstyle is included)
in order to find it.