Correct chef export usage
`chef export`'s usage was incorrect. It stated a `client.rb` would be in the root of the export and includes the `-c client.rb` in the example. `export` creates a `.chef/config.rb` and can thus be invoked with the simpler `chef-client -z`.

The message after the export command completes was correct. Now we correct it here too.