Rolling rspec-core from 3.5.1 back to 3.4.4
In audit tests we use some private rspec APIs. They changed a private
method name that we are taking advantage of causing us to get an
UndefinedMethodError. Chef does not exhibit this problem because the
last released Chef has an older version of rspec-core pinned. We see
this in he ChefDK because we are updating the ChefDK gems to take other
released changes since the last Chef release.