Prompt user when generating cookbooks with hyphens
Cookbook names with hyphens can lead to problems especially when
dealing with custom resources. We want to discourage this behavior by
informing the user that hyphenated cookbook names are discouraged and
prompts them to proceed rather than failing fast. Addtionally, the user
can override the prompt by adding the `-y` or `--yes` argument when
generating the cookbook.