These metadata fields are always helpful.
The `issues_url` and `source_url` are, indeed, helpful when the cookbook
is published to a Supermarket.  Even when they aren't, consumers of this
cookbook may want to know where the source code is stored and/or where
issues for this cookbook are tracked.

This changes updates the commentary that describes these two metadata
attributes.

Signed-off-by: Nathen Harvey <nharvey@chef.io>