Generate content with the proper yard parsable headers
The format we use here is mostly a leftover from rdoc as far as I can tell, but we’re not even consistently using that format at the moment. Fields with :: will get parsed by YARD. For a time we used Copyright:: and we seem to have slowly removed that. We should be using that format in cookbooks since it’ll get parses. Also “Cookbook Name” makes no sense when every other field omits “Name”. It should just be “Cookbook”

Signed-off-by: Tim Smith <tsmith@chef.io>