Camelizes the helper class name as well
When testing this out I used some helper names that also
needed to be camel-cased. I created a method that shares a name
with the Rails ActiveSupport method but I don't think there is a collision as that is appended onto a String and this project doesn't use that library.

Signed-off-by: Franklin Webber <franklin@chef.io>