Use `test/smoke/default` instead of `test/recipes`
Storing generated inspec tests in `test/recipes` causes Foodcritic to
fail, as Foodcritic assumes the `test` directory is now a
cookbook (which is false). Using `test/smoke/default` fixes this issue
and creates direct associations with the existing Kitchen and Workflow
templates.

1. The delivery-project.toml already runs `kitchen verify` during
smoke. There is now a connection between the phase and the tests that
are run by default.
2. The `default` folder in smoke is associated with the `default` suite
already provided in the Kitchen YAML. This indicates and guides the user
that additional suites can be created under smoke, and gives implicit
guidance as to their naming.

Signed-off-by: Tom Duffield <tom@chef.io>