Temporarily pin kitchen-dokken to 1.1.0 for 1.2 stable release
kitchen-dokken 2.x was recently released. Since this is a major version
bump, we want to give it some time to work out the kinks. We are
temporarily pinning kitchen-dokken to 1.1.0 only for the ChefDK 1.2
stable release. After that we should unpin it so that latest
kitchen-dokken ends up in ChefDK current.

Signed-off-by: Tom Duffield <tom@chef.io>