Allow to upgrade specific cookbooks in policyfile
This patches allows to update one or several cookbooks in a policyfile
instead of updating all cookbooks.

The update behavior is quite raw:
- it finds the list of cookbooks to update from command line argument
and all their dependencies
- it put hard constraints on all cookbooks except the cookbooks to
update
- it compile and install the policyfile as if we did not have a lock

This behavior is very raw since it can lead to upgrade many cookbooks at
once. We could think of adding options to control that behavior later.

This patch introduces a small change in the behavior of `-a` option for
update command:
- without this flag, attributes are updated along with cookbooks
- with this flag, only attributes are updated (no cookbook)

Fix #1182

Change-Id: I361ab141410a1bcd77e606f12d594fa830eebe0d
Signed-off-by: Grégoire Seux <g.seux@criteo.com>