remove broken/overly helpful things in version tasks
ah right, so john's thoughts on workflow were that the gem pins changed
right after release.  he was super against bumping gem versions at the
end of the release cycle because of the theory that the chef-dk release
should be "ALWAYS READY TO GO" because of $CI PHILOSOPHY.  having
accepted that there will always be manual intervention before a release
we also no longer force bumping the gems at the start of the release
cycle -- because if you're only trying to bump the minor version its
rude for the code to point a gun at your head and force you to wrangle
through bumping all our gem deps.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>