Rewrite the git source for policyfiles.
We now mutate the path source option when it is coming from an external
repo.

We also needed to suppress a Foodcritic error on the generator cookbook.
The generator cookbook does not need/contain a metadata.rb file and we
do not need to raise an error/warning for this since it is not needed.

Fixed a chefsytle issue.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>