Add new option update-strategy for policy updates
This option has two possible values for now:
- 'relaxed' (default value). This is the behavior before this patch.
  `chef update Policyfile.rb apache2` will upgrade apache2 and all its
  dependencies
- 'strict'. This behavior is very strict: it updates only the cookbook(s)
  mentionned in the command line.

The intent of this patch is to pave a possible way to have a third
strategy who would upgrade cookbooks mentionned on the command line to
their maximum existing version (possibly by upgrading other cookbooks if
necessary).

This patch already gives very fine control to users who want to update a
single cookbook (or a few cookbooks only).

Change-Id: I6ef908437ac79ea07aa0c12fb17306c359891955
Signed-off-by: Grégoire Seux <g.seux@criteo.com>