Include more information about the policy in `chef push` output
The goal of this change is to surface more information so that it's
easier for users to detect a certain class of mistakes (e.g., run
`chef push` from the wrong working directory) and to produce a more
useful log of what `chef push` did when run in a Ci context.

Possible downside is that the extra information may be offputting to
novice users, but that's probably a wash as some novices would likely
appreciate the additional detail about the consequences of the command
they are running.

Signed-off-by: Daniel DeLeo <dan@chef.io>