Branching off stable ChefDK and updating chef-apply
We branched v3.3.23 (last stable ChefDK) release and pulled in some
select changes. First, we updated chef-apply because we want users
using the latest version of that. Second, we pinned a bunch of gems
(fauxhai, inspec, etc.) that we don't want customers consuming until
the next ChefDK release. But we did allow patch fix updates and updates
to cloud gems.

We're creating a ChefDK branch for Workstation to use as we are unable
to release a stable ChefDK due to needing an updated chef-client build
which we are unable to build due to an esoteric platform requiring a new
license (this is on the way but won't be today). This branch is not
intended to be long lived and normal pinning to the stable ChefDK
release will resume once the dust settles.

Signed-off-by: tyler-ball <tball@chef.io>