restructure do_install to execute in a sub-shell
As a best practice, we /shouldn't/ change the directory inside the
callbacks since that position will be persistent throughout the build
process of the package and latter other callbacks or habitat internals
might conflict with that change of state, instead we should 1) execute
commands that depend on a specific path inside a sub-sells or 2) use
pushd and popd to move to a directory and get back where we were before.

Number 1) is more desired but you could use them both.

TODO we should define the list of binaries inside the repo and not inside
the plan

Signed-off-by: Salim Afiune <afiune@chef.io>