Adds buildtike job to test habitat package
Why? Because if we have code inside the repo that is not being tested we
will definitely loose this work while doing other modifications, the
best thing to do is always add tests to constantly check if any
functionality is lost.

TODO Improve this tests a lot since right now they are just printing the
version of the commands that work, but at least we will know if break
one for the habitat packages.

Signed-off-by: Salim Afiune <afiune@chef.io>