Bring in the non-breaking chef generate command fixes from DK 4
We cleaned up a lot of junk. A lot of it was breaking, but not all. We should bring in everything we can.  A few highlights:

A lot of readme cleanup for the generated cookbooks and repos
Fewer references to Delivery
.kitchen.yml -> kitchen.yml
knife.rb references are now config.rb
Generated files don't go into files/default anymore
Generated delivery local mode config doesn't call itself a prototype

Signed-off-by: Tim Smith <tsmith@chef.io>