Ensure we don't pull chef gem from git in chefspec
chefspec is failing verification because it's pulling `chef` from git
master.  I was not able to find what had changed, but I do see that
by specifing a no-op `GEMFILE_MOD` we'll stick with the same
chef gem version that is in the gemspec.

Applying this change specifically to the `chefspec` unit test to
prevent the env var from affecting other tests.

Fixes #2754

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>