Update project status to 'maintained'.
This sets Issue and PR SLA to 'never', and updates status to
'Maintained'.  As a product, Chef DK has been deprecated. Because we're
still releasing security updates, gem version updates, and critical fixes
the repo status of 'maintained' is more correct than 'deprecated'.

Added a description blurb to direct people to the Workstation project
and downloads.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>